Move ScrobblerCollection to trakt module
